requirements.txt
flask
torch
torchvision
transformers
Pillow
numpy